requirements.txt
Pillow==9.4.0
pystray==0.19.4
pywin32==305
screeninfo==0.8.1
six==1.16.0
